:art: Add new implementation for creating suburbs.
The new route to create suburbs sits on the Town of Interest(R)'s route.
Thus, the town model is passed to the `createSuburb` action directly. No
need to find or peek the store. Also we do away with ugly `select`
syntax that fetched all towns. It was a bad implementation from the
beginning. Super happy to take it out :v: